Major overhaul.
I updated the graphics on all of my pages to make them more visually appealing. I added a table of my skills to my About page. I streamlined all of the layouts that I had been making. I set the backgrounds to fit to the screen size that they're displaying on as well, so the site should be more mobile friendly.  Finally, I added the Projects page, where I'll keep links to all of the GitHub Project Pages that I'm working on.